+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 - Update Gradle to `8.8` (:warning: Minimal required Gradle version is `8.0`)
 
+### Dependencies
+
+infrastructure-android:
+- Remove `com.android.tools:common` from dependencies
+
 ## [0.18.1] (2023-04-18)
 
 - Fix compatibility with Gradle lower than 8.0 (#127)